Shorted the length

+++ novokhatskiy.danil/S1/inputForwardList.cpp
@@ -1,6 +1,6 @@
 #include "inputForwardList.hpp"
 
-void novokhatskiy::inputForwardList(ForwardList<std::pair<std::string, ForwardList< unsigned long long > > >& pairsOfList, std::istream& input)
+void novokhatskiy::inputForwardList(ForwardList<std::pair<std::string, ForwardList< size_t > > >& pairsOfList, std::istream& input)
 {
   std::string line = "";
   input >> line;
@@ -10,7 +10,7 @@ void novokhatskiy::inputForwardList(ForwardList<std::pair<std::string, ForwardLi
     {
       throw std::invalid_argument("Can not read the line");
     }
-    pairsOfList.push_front({ line, ForwardList< unsigned long long > ()});
+    pairsOfList.push_front({ line, ForwardList< size_t > ()});
     while (input >> line && !std::isalpha(line[0]))
     {
       pairsOfList.front().second.push_front(std::stoull(line));
